fix(reviews): address review comments + frontend lint
- parsePrProvider returns null for non-github/gitlab hosts instead of
  misclassifying everything as GitLab
- construct AiAgentReviewClassifierModel once as a field, not per-call
- drop the private validatePromptContextAccess unit test
- move ReviewEntityItem type + isReviewEntityItem out of the component
  file into reviewEntityItem.ts (fixes react-refresh only-export rule)
